Module Fun_sql

This module is used for common code across different SQL database engines. You normally will use either Fun_sqlite or Fun_postgresql as your entry- point for querying.

module type Query_sig = sig ... end

module Query : sig ... end

module type Sql = sig ... end

The definitions supported by a database engine.

module type S = sig ... end

Utilities defined in terms of the common shared interface of a database.

module Make (Sql : Sql) : S with type db = Sql.db and type arg = Sql.arg