Module Fun_sql

This module is used for common code across different SQL database engines. You normally will not be using this module, instead you would directly use either Fun_sqlite or Fun_postgresql.

module type Sql = sig ... end

module type S = sig ... end

module Make (Sql : Sql) : S with type db = Sql.db and type arg = Sql.arg and type 'a ret = 'a Sql.ret